feat: merge virtual cluster controllers
