+++ pkg/kubenest/controller/virtualcluster.node.manager/workflow/task/logger.go
